+++ packages/shared/sdk-server/src/evaluation/Evaluator.ts
@@ -17,7 +17,7 @@ import ErrorKinds from './ErrorKinds';
 import EvalResult from './EvalResult';
 import evalTargets from './evalTargets';
 import makeBigSegmentRef from './makeBigSegmentRef';
-import matchClauseWithoutSegmentOperations from './matchClause';
+import matchClauseWithoutSegmentOperations, { maybeNegate } from './matchClause';
 import matchSegmentTargets from './matchSegmentTargets';
 import { Queries } from './Queries';
 import Reasons from './Reasons';
@@ -372,10 +372,9 @@ export default class Evaluator {
             return cb(makeError(errorResult));
           }
 
-          return cb(makeMatch(match));
+          return cb(makeMatch(maybeNegate(clause, match)));
         },
       );
-      // TODO: Should this return here?
       return;
     }
     // This is after segment matching, which does not use the reference.